ukring: Make ring _alloc and _free inline functions
This commit moves the allocation and free methods of the ukring
library into the header file, making them inline.  This is done
in preparation for making the header itself global.
