+++ lib/ukring/include/uk/ring.h
@@ -368,12 +368,41 @@ uk_ring_count(struct uk_ring *br)
 			& br->br_prod_mask;
 }
 
-struct uk_ring *uk_ring_alloc(int count, struct uk_alloc *a
+static struct uk_ring *
+uk_ring_alloc(unsigned int count, size_t elemsize, struct uk_alloc *a
 #ifdef DEBUG_BUFRING
 		, struct uk_mutex *lock
 #endif
-);
-void uk_ring_free(struct uk_ring *br, struct uk_alloc *a);
+)
+{
+	struct uk_ring *br;
+
+	/* buf ring must be size power of 2 */
+	UK_ASSERT(POWER_OF_2(count));
+
+	/* Limit the size of each element to the maximum pointer size as the buffer
+	 * should not handle large elements. */
+	UK_ASSERT(elemsize <= sizeof(void *));
+
+	br = uk_malloc(a, sizeof(struct uk_ring) + count * elemsize);
+	if (br == NULL)
+		return NULL;
+#ifdef DEBUG_BUFRING
+	br->br_lock = lock;
+#endif
+	br->br_prod_size = br->br_cons_size = count;
+	br->br_prod_mask = br->br_cons_mask = count - 1;
+	br->br_prod_head = br->br_cons_head = 0;
+	br->br_prod_tail = br->br_cons_tail = 0;
+
+	return br;
+}
+
+static void
+uk_ring_free(struct uk_ring *br, struct uk_alloc *a)
+{
+	uk_free(a, br);
+}
 
 #endif
